Import Hydeout Template for thehale.github.io
I'm going to create a personal website!

After perusing lots of templates, I have decided to use the Hydeout template
(https://github.com/fongandrew/hydeout) as the base for my site.

This commit copies in the original source files of that template to provide a
clear delineation between the MIT licensed template, and my custom improvements
(which I won't be initially releasing under an open-source license).