Personalize the sidebar
This is the first commit where I really start to turn the template into my own
site. In this step, I update the `_config.yml` to use my information instead of
the Hydeout defaults. I also added my GitHub and Twitter accounts along with the
corresponding config settings/template support.

The `.gitignore` and `Gemfile.lock` were added in this stage since this stage is
when I first ran the site locally.