Refine home page layout
In the default Hydeout theme, the home page is simply a paginated list of the
blog posts.

I want to have a custom header above the list of blog posts (but only on the
first page) where I can introduce myself to visitors of my website.

This commit adds in that custom header content along with the corresponding
template changes required to meet my preferred look.