Create first post!
Now that the structure of the site is in order, it's time for some actual
blogging content!

I've decided that my first post will be about the Binary Clock app I recently
created and published on the Google Play Store.

This commit also adds a `Projects` tab to the sidebar so visitors of my site can
find information about my projects (like Binary Clock) as the number of posts
grows over time.