fix: Always show post date below title
In "Related Posts" and on category/tag pages some posts showed the dates below
the title while others had the date to the right of the title. The inconsistency
was visually distracting.

The cause was some post titles were shorter than others and the date simply
followed the title.

This commit adds a clear line break between post titles and dates so that they
always appear on separate lines.