perf: Reduce size of `barrett_delegation.jpg`
Both Cloudflare and Google have been reporting that my site takes 6-10 seconds to load, dramatically longer than I had expected.

A detailed report via Google's PageSpeed Insights pointed out this image's 4MB as a major contributor to the slowdown/bandwidth. The

The new image is down to ~74KB (a 55x reduction) making it now one of the smallest images on the site.